Suma, resta, multiplicacion y clase con metodos para ordenar y encontrar el indice en el arreglo de una posicion

+++ Par.java
@@ -2,7 +2,7 @@ public class Par {
 	int pos; // posicion
 	double valor; // magnitud
 
-	public Par (int pos, int valor){
+	public Par (int pos, double valor){
 		this.pos = pos;
 		this.valor = valor;
 	}
